requirements.txt
anyio==4.2.0
commentjson==0.9.0
grpclib==0.4.7
protobuf==4.25.2
pydantic==2.6.0
pyOpenSSL==24.0.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.31.0
xxhash==3.4.1

PyYAML~=6.0.1
aiohttp~=3.9.5